test: group BitcoinCoreIpc integration tests
Move BitcoinCoreIpc-only integration tests into bitcoin_core_ipc_integration.

Behavior changes:
- Test names and assertions stay the same.
