+++ integration-tests/tests/jd_integration.rs
@@ -1271,79 +1271,6 @@ async fn jdc_require_standard_jobs_set_does_not_group_standard_channels() {
     shutdown_all!(jdc, pool);
 }
 
-// This test verifies that JDC rejects OpenExtendedMiningChannel when
-// REQUIRES_STANDARD_JOBS is set during SetupConnection.
-#[tokio::test]
-async fn jdc_require_standard_jobs_set_rejects_open_extended_mining_channel() {
-    start_tracing();
-    let bitcoin_core = start_bitcoin_core_latest(DifficultyLevel::Low);
-    let (pool, pool_addr, jds_addr, _) =
-        start_pool_with_jds(&bitcoin_core, vec![], vec![], true).await;
-
-    let (jdc, jdc_addr, _) = start_jdc(
-        &[(pool_addr, jds_addr)],
-        ipc_config(
-            bitcoin_core.data_dir().clone(),
-            bitcoin_core.is_signet(),
-            None,
-        ),
-        vec![],
-        vec![],
-        false,
-        None,
-    );
-
-    let (sniffer, sniffer_addr) = start_sniffer("sniffer", jdc_addr, false, vec![], None);
-    // SetupConnection flags: 0b0001 == REQUIRES_STANDARD_JOBS.
-    let mock_downstream = MockDownstream::new(
-        sniffer_addr,
-        WithSetup::yes_with_defaults(Protocol::MiningProtocol, 0b0001),
-    );
-    let send_to_jdc = mock_downstream.start().await;
-
-    sniffer
-        .wait_for_message_type_and_clean_queue(
-            MessageDirection::ToDownstream,
-            MESSAGE_TYPE_SETUP_CONNECTION_SUCCESS,
-        )
-        .await;
-
-    let open_extended_mining_channel = AnyMessage::Mining(Mining::OpenExtendedMiningChannel(
-        OpenExtendedMiningChannel {
-            request_id: 100u32.into(),
-            user_identity: "user_identity".try_into().unwrap(),
-            nominal_hash_rate: 1000.0,
-            max_target: vec![0xff; 32].try_into().unwrap(),
-            min_extranonce_size: 8,
-        },
-    ));
-    send_to_jdc
-        .send(open_extended_mining_channel)
-        .await
-        .unwrap();
-
-    sniffer
-        .wait_for_message_type(
-            MessageDirection::ToDownstream,
-            MESSAGE_TYPE_OPEN_MINING_CHANNEL_ERROR,
-        )
-        .await;
-
-    let error = loop {
-        match sniffer.next_message_from_upstream() {
-            Some((_, AnyMessage::Mining(Mining::OpenMiningChannelError(msg)))) => break msg,
-            _ => continue,
-        }
-    };
-
-    assert_eq!(
-        error.error_code.as_utf8_or_hex(),
-        ERROR_CODE_OPEN_MINING_CHANNEL_EXTENDED_CHANNELS_NOT_SUPPORTED_FOR_STANDARD_JOBS
-    );
-
-    shutdown_all!(jdc, pool);
-}
-
 // Verifies that when the primary pool disconnects, JDC falls back to the next upstream entry
 // and sends *that* upstream's `user_identity` in `AllocateMiningJobToken` — not the primary's.
 #[tokio::test]
