+++ integration-tests/tests/jdc_fallback_to_solo.rs
@@ -1,92 +1,6 @@
 use integration_tests_sv2::{interceptor::MessageDirection, template_provider::DifficultyLevel, *};
 use stratum_apps::stratum_core::{common_messages_sv2::*, job_declaration_sv2::*};
 
-// launch a JDC (with Bitcoin Core IPC) connected to a Pool/JDS and then triggers a fallback to solo
-// then it mines a block using solo and verifies the block was propagated
-// meant to avoid regressions like https://github.com/stratum-mining/sv2-apps/issues/466
-#[tokio::test]
-async fn jdc_fallback_to_solo_mines_block_with_bitcoin_core_ipc() {
-    start_tracing();
-    let bitcoin_core = start_bitcoin_core_latest(DifficultyLevel::Low);
-    let current_block_hash = bitcoin_core.get_best_block_hash().unwrap();
-
-    let (pool, pool_addr, jds_addr, _) =
-        start_pool_with_jds(&bitcoin_core, vec![], vec![], false).await;
-    let (jdc_jds_sniffer, jdc_jds_sniffer_addr) = start_sniffer(
-        "jdc-fallback-bitcoin-core-jds",
-        jds_addr,
-        false,
-        vec![],
-        None,
-    );
-    let (jdc, jdc_addr, _) = start_jdc(
-        &[(pool_addr, jdc_jds_sniffer_addr)],
-        ipc_config(
-            bitcoin_core.data_dir().clone(),
-            bitcoin_core.is_signet(),
-            None,
-        ),
-        vec![],
-        vec![],
-        false,
-        None,
-    );
-
-    // assert JDC-JDS connection is established
-    {
-        jdc_jds_sniffer
-            .wait_for_message_type(MessageDirection::ToUpstream, MESSAGE_TYPE_SETUP_CONNECTION)
-            .await;
-        jdc_jds_sniffer
-            .wait_for_message_type(
-                MessageDirection::ToDownstream,
-                MESSAGE_TYPE_SETUP_CONNECTION_SUCCESS,
-            )
-            .await;
-        jdc_jds_sniffer
-            .wait_for_message_type(
-                MessageDirection::ToUpstream,
-                MESSAGE_TYPE_ALLOCATE_MINING_JOB_TOKEN,
-            )
-            .await;
-        jdc_jds_sniffer
-            .wait_for_message_type(
-                MessageDirection::ToDownstream,
-                MESSAGE_TYPE_ALLOCATE_MINING_JOB_TOKEN_SUCCESS,
-            )
-            .await;
-    }
-
-    // trigger JDC fallback
-    pool.shutdown().await;
-    tokio::time::sleep(tokio::time::Duration::from_secs(2)).await;
-
-    let (tproxy, tproxy_addr, _) =
-        start_sv2_translator(&[jdc_addr], false, vec![], vec![], None, false).await;
-    let (_minerd_process, _minerd_addr) = start_minerd(tproxy_addr, None, None, false).await;
-
-    let timeout = tokio::time::Duration::from_secs(60);
-    let poll_interval = tokio::time::Duration::from_secs(2);
-    let start_time = tokio::time::Instant::now();
-
-    // assert JDC was able to propagate a block while doing solo
-    loop {
-        tokio::time::sleep(poll_interval).await;
-        let new_block_hash = bitcoin_core.get_best_block_hash().unwrap();
-        if new_block_hash != current_block_hash {
-            shutdown_all!(jdc, tproxy);
-            return;
-        }
-        if start_time.elapsed() > timeout {
-            panic!(
-                "JDC fallback to solo with BitcoinCoreIpc should have propagated a new block \
-                 within {} seconds",
-                timeout.as_secs()
-            );
-        }
-    }
-}
-
 // launch a JDC (with Sv2 TP over TCP) connected to a Pool/JDS and then triggers a fallback to solo
 // then it mines a block using solo and verifies the block was propagated
 // meant to avoid regressions like https://github.com/stratum-mining/sv2-apps/issues/466
