+++ integration-tests/tests/pool_integration.rs
@@ -1029,72 +1029,3 @@ async fn pool_require_standard_jobs_set_does_not_group_standard_channels() {
     }
     pool.shutdown().await;
 }
-
-// This test verifies that Pool rejects OpenExtendedMiningChannel when
-// REQUIRES_STANDARD_JOBS is set during SetupConnection.
-#[tokio::test]
-async fn pool_require_standard_jobs_set_rejects_open_extended_mining_channel() {
-    start_tracing();
-    let bitcoin_core = start_bitcoin_core_latest(DifficultyLevel::Low);
-    let (pool, pool_addr, _) = start_pool(
-        ipc_config(
-            bitcoin_core.data_dir().clone(),
-            bitcoin_core.is_signet(),
-            None,
-        ),
-        vec![],
-        vec![],
-        false,
-    )
-    .await;
-
-    let (sniffer, sniffer_addr) = start_sniffer("sniffer", pool_addr, false, vec![], None);
-    // SetupConnection flags: 0b0001 == REQUIRES_STANDARD_JOBS.
-    let mock_downstream = MockDownstream::new(
-        sniffer_addr,
-        WithSetup::yes_with_defaults(Protocol::MiningProtocol, 0b0001),
-    );
-    let send_to_pool = mock_downstream.start().await;
-
-    sniffer
-        .wait_for_message_type_and_clean_queue(
-            MessageDirection::ToDownstream,
-            MESSAGE_TYPE_SETUP_CONNECTION_SUCCESS,
-        )
-        .await;
-
-    let open_extended_mining_channel = AnyMessage::Mining(Mining::OpenExtendedMiningChannel(
-        OpenExtendedMiningChannel {
-            request_id: 100u32.into(),
-            user_identity: "user_identity".try_into().unwrap(),
-            nominal_hash_rate: 1000.0,
-            max_target: vec![0xff; 32].try_into().unwrap(),
-            min_extranonce_size: 8,
-        },
-    ));
-    send_to_pool
-        .send(open_extended_mining_channel)
-        .await
-        .unwrap();
-
-    sniffer
-        .wait_for_message_type(
-            MessageDirection::ToDownstream,
-            MESSAGE_TYPE_OPEN_MINING_CHANNEL_ERROR,
-        )
-        .await;
-
-    let error = loop {
-        match sniffer.next_message_from_upstream() {
-            Some((_, AnyMessage::Mining(Mining::OpenMiningChannelError(msg)))) => break msg,
-            _ => continue,
-        }
-    };
-
-    assert_eq!(
-        error.error_code.as_utf8_or_hex(),
-        ERROR_CODE_OPEN_MINING_CHANNEL_EXTENDED_CHANNELS_NOT_SUPPORTED_FOR_STANDARD_JOBS
-    );
-
-    pool.shutdown().await;
-}
